Hrchanges

+++ LMS_CMS_Backend/LMS_CMS_BL/DTO/HR/BounsGetDTO.cs
@@ -17,7 +17,7 @@ namespace LMS_CMS_BL.DTO.HR
         public int NumberOfBounsDays { get; set; }
         public int Hours { get; set; }
         public int Minutes { get; set; }
-        public long Amount { get; set; }
+        public decimal Amount { get; set; }
         public string? Notes { get; set; }
         public long EmployeeID { get; set; }
         public string EmployeeEnName { get; set; }
